move full-history-backfill subcommand factory into backfill package
- Relocate NewCmd() from cmd/stellar-rpc/main.go into
  cmd/stellar-rpc/internal/fullhistory/backfill/cmd.go so subcommand-
  specific wiring lives with the package whose body it will drive in
  later slices; main.go shrinks to a single AddCommand call.
- Move the factory test alongside as backfill/cmd_test.go.
- Drop the backfill/doc.go placeholder (cmd.go now populates the
  package).
- Address Copilot feedback on #699:
  * workers flag description was misleading (claimed default GOMAXPROCS
    while literal default was 0); rewrite as "(0 = GOMAXPROCS)" to
    match the sentinel convention.
  * MarkFlagRequired("config") error is now explicitly handled via
    panic — can only fire on a typo in a future refactor, and a silent
    drop would hide the break.
