--- cmd/stellar-rpc/internal/fullhistory/backfill/doc.go
@@ -1,5 +1,0 @@
-// Package backfill implements the offline full-history backfill pipeline.
-//
-// This package is currently a placeholder so the package tree compiles; the
-// pipeline implementation lands in subsequent slices.
-package backfill
