+++ cmd/stellar-rpc/main.go
@@ -11,28 +11,8 @@ import (
 
 	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/config"
 	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/daemon"
+	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/fullhistory/backfill"
 )
-
-// newFullHistoryBackfillCmd builds the `full-history-backfill` subcommand,
-// an offline ingest entry point whose body is wired up in subsequent slices.
-func newFullHistoryBackfillCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "full-history-backfill",
-		Short: "Offline backfill of historical Stellar ledger data",
-		RunE: func(cmd *cobra.Command, _ []string) error {
-			_, err := fmt.Fprintln(cmd.OutOrStdout(), "full-history-backfill: not yet implemented")
-			return err
-		},
-	}
-	cmd.Flags().String("config", "", "Path to TOML configuration file (required)")
-	cmd.Flags().Uint32("start-ledger", 0, "First ledger to ingest (inclusive, >= 2)")
-	cmd.Flags().Uint32("end-ledger", 0, "Last ledger to ingest (inclusive, > start-ledger)")
-	cmd.Flags().Int("workers", 0, "Concurrent DAG task slots (default GOMAXPROCS)")
-	cmd.Flags().Int("max-retries", 3, "Max retries per task before marking failed")
-	cmd.Flags().Bool("verify-recsplit", true, "Run RecSplit verify phase after build")
-	_ = cmd.MarkFlagRequired("config")
-	return cmd
-}
 
 func main() {
 	var cfg config.Config
@@ -102,7 +82,7 @@ func main() {
 
 	rootCmd.AddCommand(versionCmd)
 	rootCmd.AddCommand(genConfigFileCmd)
-	rootCmd.AddCommand(newFullHistoryBackfillCmd())
+	rootCmd.AddCommand(backfill.NewCmd())
 
 	if err := cfg.AddFlags(rootCmd); err != nil {
 		fmt.Fprintf(os.Stderr, "could not parse config options: %v\n", err)
